Plugin names and discovery methods corrected.
- Plugin name is changed to entity limit.
- Plugin discovery method is added.
- Use plugin discovery to select plugin in add form.
