+++ src/Plugin/EntityLimitPluginInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\entity_limit\Plugin;
+
+use Drupal\Component\Plugin\PluginInspectionInterface;
+use Drupal\Core\Plugin\PluginFormInterface;
+use Drupal\Core\Executable\ExecutableInterface;
+
+/**
+ * Interface for plugins of entity limit.
+ */
+interface EntityLimitPluginInterface extends PluginFormInterface, PluginInspectionInterface, ExecutableInterface {
+  /**
+   * @TODO: Add methods as and when needed.
+   */
+
+}
